Refactor menu code and add tests.
